Update README.md
The main menu bar contains the same menu strip items as the Standard Windows Calculator. 
The event handlers are added to handle the Click event for Edit/Copy,  Edit/Paste, and Help/About menu strip items.
For the remaining menu items create handlers that will display a message box with the “Will be developed!” message. 
In addition, the context menus for all memory buttons are designed in order to help the user to understand the operation of each memory button (i.e., MC, MR, MS, M+, and M-).
The calculator also prevents run-time errors such as a division by 0 and a square root of a negative number.